Resolve some review comments
