+++ elide-async/src/main/java/com/yahoo/elide/async/models/AsyncQuery.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.async.models;
 
-import java.util.Date;
 import java.util.UUID;
 
 import javax.inject.Inject;
@@ -13,8 +12,6 @@ import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
 import javax.persistence.Transient;
 
 import com.yahoo.elide.annotation.DeletePermission;
@@ -38,7 +35,7 @@ import lombok.Data;
 @UpdatePermission(expression = "Prefab.Role.None")
 @DeletePermission(expression = "Prefab.Role.None")
 @Data
-public class AsyncQuery implements PrincipalOwned {
+public class AsyncQuery extends AsyncBase implements PrincipalOwned {
     @Id
     private UUID id; //Can be generated or provided.
 
@@ -52,10 +49,6 @@ public class AsyncQuery implements PrincipalOwned {
     @OneToOne(mappedBy = "query", cascade = CascadeType.REMOVE)
     private AsyncQueryResult result;
 
-    private Date createdOn;
-
-    private Date updatedOn;
-
     @Inject
     @Transient
     private AsyncExecutorService asyncExecutorService;
@@ -63,22 +56,9 @@ public class AsyncQuery implements PrincipalOwned {
     @Exclude
     private String principalName;
 
-    @Exclude
-    protected String naturalKey = UUID.randomUUID().toString();
-
     @Override
     public String getPrincipalName() {
         return principalName;
-    }
-
-    @PrePersist
-    public void prePersist() {
-        createdOn = updatedOn = new Date();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        this.updatedOn = new Date();
     }
 
     @OnCreatePreSecurity
@@ -91,17 +71,4 @@ public class AsyncQuery implements PrincipalOwned {
         asyncExecutorService.executeQuery(this, scope.getUser());
     }
 
-    @Override
-    public int hashCode() {
-        return naturalKey.hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null || !(obj instanceof AsyncQuery)) {
-            return false;
-        }
-
-        return ((AsyncQuery) obj).naturalKey.equals(naturalKey);
-    }
 }
