+++ elide-async/src/main/java/com/yahoo/elide/async/models/AsyncQueryResult.java
@@ -5,14 +5,11 @@
  */
 package com.yahoo.elide.async.models;
 
-import java.util.Date;
 import java.util.UUID;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
 
 import com.yahoo.elide.annotation.CreatePermission;
 import com.yahoo.elide.annotation.DeletePermission;
@@ -33,7 +30,7 @@ import lombok.Data;
 @CreatePermission(expression = "Prefab.Role.None")
 @DeletePermission(expression = "Prefab.Role.None")
 @Data
-public class AsyncQueryResult implements PrincipalOwned {
+public class AsyncQueryResult extends AsyncBase implements PrincipalOwned {
     @Id
     private UUID id; //Matches UUID in query.
 
@@ -43,42 +40,12 @@ public class AsyncQueryResult implements PrincipalOwned {
 
     private Integer status; // HTTP Status
 
-    private Date createdOn;
-
-    private Date updatedOn;
-
     @OneToOne
     private AsyncQuery query;
-
-    @Exclude
-    protected String naturalKey = UUID.randomUUID().toString();
 
     @Exclude
     public String getPrincipalName() {
         return query.getPrincipalName();
     }
 
-    @PrePersist
-    public void prePersist() {
-        createdOn = updatedOn = new Date();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        updatedOn = new Date();
-    }
-
-    @Override
-    public int hashCode() {
-        return naturalKey.hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null || !(obj instanceof AsyncQuery)) {
-            return false;
-        }
-
-        return ((AsyncQuery) obj).naturalKey.equals(naturalKey);
-    }
 }
